\name{lpacf-package}
\alias{lpacf-package}
\docType{package}
\title{Compute localized partial autocorrelation for locally stationary wavelet time series.
}
\description{This package computes a localized partial autocorrelation of a time series assuming a locally stationary wavelet model.}
\details{
\tabular{ll}{
Package: \tab lpacf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2023-03-07\cr
License: \tab GPL-2\cr
}
The \code{\link{lpacf}} function computes the localized partial autocorrelation
of a locally stationary time series. The resultant object can be printed using
\code{\link{print.lpacf}} or plotted by \code{\link{plot.lpacf}}.
}
\author{
Rebecca Killick, Marina Knight, Guy Nason and Matt Nunes

Maintainer: Rebecca Killick <r.killick@lancs.ac.uk>
}
\references{
Killick, R., Knight, M. I., Nason, G.P. and Eckley, I. A. (2020) 
The local partial autocorrelation function and some applications. 
\emph{Electron. J. Statist.} \bold{14} (2), 3268-3314. 
DOI: \href{https://www.doi.org/}{10.1214/20-EJS1748}.
}
\keyword{ package }
\seealso{
	\code{\link{lpacf}}
}
\examples{
#
# See examples in each of the functions' help pages linked above.
#
}
